Extract Spring MVC support into separate mvn module
