+++ pom.xml
@@ -64,10 +64,7 @@
 
     <properties>
         <jdk.version>1.8</jdk.version>
-        <spring.version>4.1.6.RELEASE</spring.version>
         <slf4j.version>1.7.5</slf4j.version>
-        <jcl.slf4j.version>1.7.12</jcl.slf4j.version>
-        <jackson.version>2.6.0</jackson.version>
 
         <github-repo-owner>excelsior-oss</github-repo-owner>
         <github-repo>restler</github-repo>
@@ -88,6 +85,7 @@
             </activation>
             <modules>
                 <module>restler-core</module>
+                <module>restler-spring-mvc</module>
                 <module>restler-integration-tests</module>
             </modules>
         </profile>
@@ -96,6 +94,7 @@
             <!-- Due to the skip-deployment bug in the Nexus Staging plugin, we have to exclude the module to not deploy it  -->
             <modules>
                 <module>restler-core</module>
+                <module>restler-spring-mvc</module>
             </modules>
             <build>
                 <plugins>
@@ -219,7 +218,14 @@
             <artifactId>cglib-nodep</artifactId>
             <version>3.1</version>
         </dependency>
-        <!-- Groovy and Spock are for testing -->
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>${slf4j.version}</version>
+        </dependency>
+
+        <!-- Test dependencies -->
         <dependency>
             <groupId>org.codehaus.groovy.maven.runtime</groupId>
             <artifactId>gmaven-runtime-api</artifactId>
@@ -244,8 +250,6 @@
             <version>1.0-groovy-2.4</version>
             <scope>test</scope>
         </dependency>
-
-        <!-- Required by Spock -->
         <dependency>
             <groupId>org.objenesis</groupId>
             <artifactId>objenesis</artifactId>
@@ -253,41 +257,5 @@
             <scope>test</scope>
         </dependency>
 
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-core</artifactId>
-            <version>${spring.version}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>jcl-over-slf4j</artifactId>
-            <version>${jcl.slf4j.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-web</artifactId>
-            <version>${spring.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-            <version>${slf4j.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>${jackson.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-            <version>1.1.3</version>
-        </dependency>
     </dependencies>
 </project>
