+++ restler-integration-tests/pom.xml
@@ -13,6 +13,9 @@
     <properties>
         <kotlin.version>0.12.200</kotlin.version>
         <spring.security.verion>3.2.7.RELEASE</spring.security.verion>
+        <spring.version>4.1.6.RELEASE</spring.version>
+        <jcl.slf4j.version>1.7.12</jcl.slf4j.version>
+        <jackson.version>2.6.0</jackson.version>
     </properties>
 
     <artifactId>test-server</artifactId>
@@ -81,12 +84,14 @@
     </build>
 
     <dependencies>
+
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>restler-core</artifactId>
+            <artifactId>restler-spring-mvc</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
         </dependency>
+
 
         <dependency>
             <groupId>org.springframework</groupId>
@@ -103,16 +108,16 @@
             <artifactId>spring-security-config</artifactId>
             <version>${spring.security.verion}</version>
         </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.module</groupId>
+            <artifactId>jackson-module-paranamer</artifactId>
+            <version>${jackson.version}</version>
+        </dependency>
 
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
             <artifactId>kotlin-stdlib</artifactId>
             <version>${kotlin.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.module</groupId>
-            <artifactId>jackson-module-paranamer</artifactId>
-            <version>${jackson.version}</version>
         </dependency>
 
         <dependency>
@@ -121,12 +126,35 @@
             <version>9.3.2.v20150730</version>
         </dependency>
 
+        <!-- Setup logging with slf4j -->
+        <dependency>
+            <groupId>ch.qos.logback</groupId>
+            <artifactId>logback-classic</artifactId>
+            <version>1.1.3</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-core</artifactId>
+            <version>${spring.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>commons-logging</groupId>
+                    <artifactId>commons-logging</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>jcl-over-slf4j</artifactId>
+            <version>${jcl.slf4j.version}</version>
+        </dependency>
         <!-- required for logback groovy config -->
         <dependency>
             <groupId>org.codehaus.groovy</groupId>
             <artifactId>groovy-all</artifactId>
             <version>2.4.3</version>
         </dependency>
+
     </dependencies>
 
 </project>
