+++ restler-spring-mvc/src/main/java/org/restler/spring/mvc/HttpCallTypeReference.java
@@ -1,4 +1,4 @@
-package org.restler.spring;
+package org.restler.spring.mvc;
 
 import org.restler.client.Call;
 import org.springframework.core.ParameterizedTypeReference;
