+++ restler-spring-mvc/src/main/java/org/restler/spring/mvc/ParameterResolver.java
@@ -1,4 +1,4 @@
-package org.restler.client;
+package org.restler.spring.mvc;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
@@ -7,10 +7,10 @@ import java.util.Optional;
 @FunctionalInterface
 public interface ParameterResolver {
 
-    Optional<String> resolve(Method m, Object[] args, Annotation[][] annotations, String[] paramNames, int paramIdx);
-
     static ParameterResolver valueOfParamResolver() {
         return (m, args, annotations, paramNames, paramIdx) -> Optional.ofNullable(args[paramIdx]).map(String::valueOf);
     }
+
+    Optional<String> resolve(Method m, Object[] args, Annotation[][] annotations, String[] paramNames, int paramIdx);
 
 }
