+++ restler-spring-mvc/src/test/java/org/restler/spring/mvc/Greeter.java
@@ -1,4 +1,4 @@
-package org.restler.spring;
+package org.restler.spring.mvc;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
